changed the clickhouse scema datatype

+++ client/pkg/clickhouse/statements.go
@@ -153,7 +153,7 @@ const azureContainerPushEventTable DBStatement = `
 		ImageName String,
 		Event String,
 		Timestamp String,
-		Size UInt64,
+		Size Int32,
 		SHAID String
 	) engine=File(TabSeparated)
 	`
